Developer Documentation

This documentation describes the internal functioning of the library. For most users, the User Manual is sufficient.

Full GRAPE

The basic GRAPE implementation works on the full pulse and uses tracking to refocus the propagator at regular time intervals. Some mathematical details are given in Background Details.

UTrack.FullPulse — Type

FullPulse

Type of a full pulse. Currently just an alias for Vector{Float64}.

The state of the optimization is stored in an OptimizationState, which automatically detects if CUDA is available and falls back to a CPU implementation otherwise.

UTrack.OptimizationState — Type

OptimizationState{T <: AbstractArray{Float64}, S <: AbstractArray{ComplexF64}}

Stores the state for the full GRAPE optimization. Uses CUDA if CUDA.functional() == true.

UTrack.OptimizationState — Method

OptimizationState(p::OptimizationParameters)

Create an empty state according to the given parameters.

UTrack.set_pulse — Method

set_pulse(s::OptimizationState{T,S}, ϕ::FullPulse) where {T<: CuArray, S <: CuArray}

Assign the pulse ϕ to the state s. Automatically converts the pulse to a CuArray.

UTrack.set_pulse — Method

set_pulse(s::OptimizationState{T,S}, ϕ::FullPulse) where {T<: Array, S <: Array}

Assign the pulse ϕ to the state s.

UTrack.get_gradient — Method

get_gradient(s::OptimizationState, δϕ)

Store the gradient from state s in δϕ.

Some information can be computed in advance once the OptimizationParameters are known.

UTrack.OptimizationPrecomputed — Type

OptimizationPrecomputed{T <: AbstractArray{Float64}, S <: AbstractArray{ComplexF64}}

Stores some information that only has to be computed once and can be shared among all optimizations using the same parameters. Uses CUDA if available.

UTrack.OptimizationPrecomputed — Method

OptimizationPrecomputed(p::OptimizationParameters)

Precomputes some information for parameters p.

The GRAPE algorithm is performed by the following two functions which execute CUDA kernels (if available).

UTrack.propagate_state_compute_cost — Function

propagate_state_compute_cost(p::OptimizationParameters, s::OptimizationState, pc::OptimizationPrecomputed)

Initializes and propagates the state, and computes the cost function at the same time. Returns the value of the cost function. Uses CUDA if available.

UTrack.backpropagate_compute_gradient — Function

backpropagate_compute_gradient(p::OptimizationParameters, s::OptimizationState, pc::OptimizationPrecomputed)

Backpropagates the state and computes the gradient stored in s. Only call this after propagate_state_compute_cost. Uses CUDA if available.

Cycle GRAPE (XUR)

UTrack.XUR_Pulse — Type

mutable struct XUR_Pulse

Type of an extended UR (XUR) pulse.

The state of the optimization is stored in an OptimizationStateXUR object.

UTrack.OptimizationStateXUR — Type

OptimizationStateXUR {T <: AbstractArray{Float64}, S <: AbstractArray{ComplexF64}, R <: AbstractArray{Int64}}

Stores the state for the cycle GRAPE optimization.

UTrack.OptimizationStateXUR — Method

OptimizationStateXUR(p::OptimizationParameters)

Create an empty state according to the given parameters.

UTrack.set_pulse — Method

set_pulse(s::OptimizationStateXUR{T,S,R}, Φ) where {T<: CuArray, S <: CuArray, R <: CuArray}

Assign the pulse phases Φ to the state s. Automatically converts the pulse phases to a CuArray.

UTrack.set_pulse — Method

set_pulse(s::OptimizationStateXUR{T,S,R}, Φ) where {T<: Array, S <: Array, R <: Array}

Assign the pulse phases Φ to the state s.

UTrack.set_types — Function

set_types(s::OptimizationStateXUR{T,S,R}, τ) where {T<: CuArray, S <: CuArray, R <: CuArray}

Sets the UR cycle types of the state s. Automatically converts to a CuArray.

set_types(s::OptimizationStateXUR{T,S,R}, τ) where {T<: Array, S <: Array, R <: Array}

Sets the UR cycle types of the state s.

UTrack.get_gradient — Method

get_gradient(s::OptimizationStateXUR, δΦ)

Store the gradient from state s in δΦ.

As in the full optimization, the precomputed data is stored in an OptimizationPrecomputedXUR object.

UTrack.OptimizationPrecomputedXUR — Type

struct OptimizationPrecomputedXUR{T <: AbstractArray{Float64}, S <: AbstractArray{ComplexF64}}

Stores some information that only has to be computed once and can be shared among all optimizations using the same parameters. Uses CUDA if available.

UTrack.OptimizationPrecomputedXUR — Method

OptimizationPrecomputedXUR(p::OptimizationParameters)

Precomputes some information for parameters p for the cycle optimization.

The cycle GRAPE algorithm is performed by the following two functions which execute CUDA kernels if available.

UTrack.propagate_state_compute_cost_xur — Function

propagate_state_compute_cost_xur(p::OptimizationParameters, s::OptimizationStateXUR, pc::OptimizationPrecomputedXUR)

Initializes and propagates the state, and computes the cost function at the same time. Returns the value of the cost function.

UTrack.backpropagate_compute_gradient_xur — Function

backpropagate_compute_gradient_xur(p::OptimizationParameters, s::OptimizationStateXUR, pc::OptimizationPrecomputedXUR)

Backpropagates the state and computes the gradient stored in s. Only call this after propagate_state_compute_cost_xur.

Gradient Optimization

For the gradient based optimization of the pulse, we combine the GRAPE algorithm to compute the gradient efficiently with the general purpose optimization package Optim. In particular we rely on the optimization algorithms BFGS and L-BFGS.

A list of pulses with the same parameters can be optimized simultaneously.

UTrack.optimize_pulses! — Function

optimize_pulses!(p::OptimizationParameters, pulses; print_func=nothing, traces=nothing)

Optimizes a collection of pulses. Can be FullPulses or XUR_Pulses. Sorts pulses according to increasing cost function value.

A single optimization is implemented using the following functions.

UTrack.fg_full! — Function

fg_full!(F, G, ϕ, p, s, pc)

Evaluates cost function and gradient for use in Optim.optimize

UTrack.fg_xur! — Function

fg_xur!(F, G, Φ, p, s, pc)

Evaluates cost function and gradient for use in Optim.optimize

UTrack.optimize_pulse — Function

optimize_pulse(p::OptimizationParameters, pc::OptimizationPrecomputed, pulse::FullPulse, method, opts)

Optimizes a pulse, returns full optimization result.

optimize_pulse(p::UTrack.OptimizationParameters, pc::UTrack.OptimizationPrecomputedXUR, xur::UTrack.XUR_Pulse, method, opts)

Optimizes an XUR pulse, returns full optimization result.

Genetic Algorithm

To avoid getting stuck in local optima, the gradient based optimization is complemented with a genetic algorithm to explore larger regions of the optimization space.

UTrack.evolve_generation — Function

evolve_generation(p::OptimizationParameters, pulses::Vector{UTrack.FullPulse})

A list of pulses (ordered from best to worst) is modified via genetic update and the list of modified pulses is returned. The best pulse is always preserved (elitism). The new pulses are obtained via recombination or permutation.

evolve_generation(p::OptimizationParameters, pulses::Vector{UTrack.XUR_Pulse})

A list of XUR pulses (ordered from best to worst) is modified via genetic update and the list of modified pulses is returned. The best pulse is always preserved (elitism). The new pulses are obtained via recombination or permutation.

Complete Optimization

As described in User Manual, in practice it is recommended to use new_optimization and continue_optimization in practice.

UTrack.perform_optimization_steps! — Function

perform_optimization_steps!(p::OptimizationParameters, pulses, folder::AbstractString, latest=0)

Optimizes a list of pulses. Saves pulses and plots to folder. The pulses can be of type FullPulse or XUR_Pulse.

UTrack.new_optimization — Function

new_optimization(folder::AbstractString, params::OptimizationParameters)

Starts new optimization based on OptimizationParameters params in folder. The folder will be created, so it may not exist, but its direct parent folder has to exist, see Base.Filesystem.mkdir.

UTrack.new_optimization_no_xur — Function

new_optimization_no_xur(folder::AbstractString, p::OptimizationParameters)

Starts new optimization based on OptimizationParameters params in folder. This function only uses full pulse optimization and no XUR. It is recommended to use new_optimization in practice for better performance. The folder will be created, so it may not exist, but the direct parent folder has to exist, see Base.Filesystem.mkdir.

UTrack.continue_optimization — Function

continue_optimization(folder::AbstractString)

Continue optimization started in folder. Use this if the optimization started using new_optimization was interrupted.

Persistence Functions

UTrack.setup_folder — Method

setup_folder(folder::AbstractString, p::OptimizationParameters)

Creates folder structure in folder to store all output related to the optimization as in the following example:

folder
├── parameters.json
├── plots
│   ├── final
│   │   ├── cost-profile.pdf
│   │   ├── full-pulse.pdf
│   │   └── pulse-cycles
│   │       ├── cycle-1.pdf
│   │       ├── ...
│   │       └── cycle-10.pdf
│   ├── optimization-cycle
│   │   ├── evolution.pdf
│   │   └── generations
│   │       ├── gen-1.pdf
│   │       ├── ...
│   │       └── gen-20.pdf
│   └── optimization-full
│       ├── evolution.pdf
│       └── generations
│           ├── gen-1.pdf
│           ├── ...
│           └── gen-15.pdf
└── pulses
    ├── final
    │   └── optimal-pulse.mat
    ├── optimization-cycle
    │   └── generations
    │       ├── gen-0.csv
    │       ├── ...
    │       └── gen-20.csv
    └── optimization-full
        └── generations
            ├── gen-0.csv
            ├── ...
            └── gen-15.csv

UTrack.save_parameters — Method

save_parameters(folder::AbstractString, p::OptimizationParameters)

Store the optimization parameters p in the folder.

UTrack.load_parameters — Method

load_parameters(folder::AbstractString)::OptimizationParameters

Read and return the optimization parameters stored in the folder.

UTrack.save_generation — Method

save_generation(folder::AbstractString, gen::Integer, p::OptimizationParameters, pulses::AbstractArray{FullPulse})

Save all pulses in a CSV file.

UTrack.save_generation — Method

save_generation(folder::AbstractString, gen::Integer, p::OptimizationParameters, xurs::AbstractArray{XUR_Pulse})

Save all pulses in a CSV file. The XUR pulses are stored as full pulses.

UTrack.load_generation — Method

load_generation(folder::AbstractString, gen::Integer, p::OptimizationParameters, type)

Load pulses from generation gen of given type, where type can be UTrack.FullPulse or UTrack.XUR_Pulse.

UTrack.latest_gen — Method

latest_gen(folder::AbstractString, type=UTrack.FullPulse)

Return index of latest generation in folder of given type (starting at 0). Returns -1 if empty.

UTrack.load_latest_generation — Method

load_latest_generation(folder::AbstractString)

Load the latest generation in folder overall.

UTrack.load_best_pulse — Method

load_best_pulse(folder::AbstractString)

Load best pulse from latest generation from folder.

UTrack.plot_optimization — Function

plot_optimization(folder::AbstractString, filename::AbstractString, type)

Plot the costs of all pulses of the given type over all generations.

UTrack.plot_grape_convergence — Function

plot_grape_convergence(folder::AbstractString, traces, gen::Integer, type)

Plot the traces of the GRAPE optimizations in generation gen for a given type.

UTrack.plot_and_write_final — Function

plot_and_write_final(p::OptimizationParameters, folder::AbstractString)

Create all remaining plots and output files after the optimization is done.

Helper Functions

Printing

UTrack.trace_printing — Method

trace_printing(x)

Print single iteration of GRAPE optimization. Use as follows: Optim.Options(..., show_trace=false, callback=trace_printing).

UTrack.generation_printing — Method

generation_printing(gen::Integer, values)

Print single generation of genetic algorithm.

Plotting

UTrack.plot_pulse — Method

plot_pulse(ϕ, filename::AbstractString)

Plot a pulse ϕ to file.

UTrack.plot_pulsecycles — Method

plot_pulsecycles(p::OptimizationParameters, ϕ, folder::AbstractString)

Plot each cycle separately in file "$folder/cycle-$i.pdf".

UTrack.plot_many_convergence — Method

plot_many_convergence(traces, filename::AbstractString)

Plot convergence traces of several optimizations.

UTrack.plot_cost_profile — Function

plot_cost_profile(p::OptimizationParameters, ϕ::FullPulse, filename::AbstractString)

Plot cost profile of the pulse ϕ.

I/O

The following functions are used for reading and writing parameters and pulses from and to files.

Parameters are stored in JSON files.

UTrack.read_parameters — Method

read_parameters(filename::AbstractString)

Read optimization parameters from JSON file.

UTrack.write_parameters — Method

write_parameters(filename::AbstractString, p::OptimizationParameters)

Write optimization parameters to JSON file.

Pulses (of the same length) are stored in CSV files, one pulse per column.

UTrack.read_all_pulses_full — Method

read_all_pulses_full(filename::AbstractString)

Read pulses from CSV file.

UTrack.read_all_pulses_xur — Method

read_all_pulses_xur(filename::AbstractString, p::OptimizationParameters)

Read pulses from CSV file and return as XUR pulses.

UTrack.write_all_pulses_full — Method

write_all_pulses_full(filename::AbstractString, pulses::AbstractArray{FullPulse})

Write a list of pulses of equal length to a CSV file.

UTrack.write_all_pulses_xur — Method

write_all_pulses_xur(filename::String, p::OptimizationParameters, xur_pulses::AbstractArray{XUR_Pulse})

Write a list of XUR pulses (converted to full pulses) of equal length to a CSV file.

For convenience, a single pulse can be stored in a Matlab file.

UTrack.read_pulse_matlab — Function

read_pulse_matlab(filename::AbstractString)

Read single pulse from MATLAB file.

UTrack.write_pulse_matlab — Function

write_pulse_matlab(filename::AbstractString, pulse::FullPulse)

Write single pulse to MATLAB file. The variable name is set to phi.

Pulses

UTrack.UR_cycle — Function

UR_cycle(n::Integer, ϕ2::Float=nothing, sign::Integer=1)

Computes the UR pulse sequence (single cycle) of length n with first phase zero, second phase equal to 2πk/n, and sign given by sign. We require k=0,...,n-1, and s=±1.

UTrack.UR_cycles — Function

UR_cycles(n::Integer)

Returns a Matrix of size n x 2n containing all UR_n pulses as columns.

UTrack.PulseType — Type

@enum PulseType

Pulse types used for initialization.

UTrack.UR_pulse — Function

UR_pulse(p::OptimizationParameters, ϕ2::Float=0, sign::Integer=1)

Computes a periodic UR pulse sequence with cycle length p.NSteps and number of cycles p.NCycles. The remaining arguments are as in UR_cycle.

UTrack.initialize_pulse — Function

initialize_pulse(p::OptimizationParameters, type::PulseType)

Initialize a single pulse of given pulse type according to the given parameters.

UTrack.initialize_pulses — Function

initialize_pulses(p::OptimizationParameters, type::PulseType=Rand)

Initialize population of pulses of given pulse type according to the given parameters.

UTrack.renormalize_pulses! — Function

renormalize_pulses!(pulses::AbstractArray{FullPulse})

Normalizes an array of full pulses to have phases in [0,2π).

renormalize_pulses!(pulses::AbstractArray{XUR_Pulse})

Normalizes an array of XUR pulses to have phases in [0,2π).